refactor: move data to services
