+++ src/screens/Collections/index.int.spec.tsx
@@ -6,7 +6,7 @@ import { Text } from 'react-native'
 
 import * as factories from '../../../test/factories'
 import withContexts from '../../components/with-contexts'
-import * as collections from '../../data/collections'
+import * as collections from '../../services/data/collections'
 import { CollectionData } from '../../types/data'
 import Screens, { AppStackParams } from '../screens'
 import { collectionsScreen } from '.'
