+++ src/services/data/bookmarks.ts
@@ -1,15 +1,16 @@
-import Languages from '../lib/languages'
-import { ShabadData } from '../types/data'
+import Languages from '../../lib/languages'
+import { BookmarkData } from '../../types/data'
 import * as gurbaniNow from './gurbaninow'
 
-export const getShabad = async ( id: string ): Promise<ShabadData> => {
+export const getBookmark = async ( id: string ): Promise<BookmarkData> => {
   const {
-    shabadinfo: { source, writer, pageno },
-    shabad,
-  } = await gurbaniNow.getShabad( id )
+    baniinfo: { source, writer, akhar },
+    bani,
+  } = await gurbaniNow.getBani( id )
 
   return {
     id,
+    nameGurmukhi: akhar,
     source: {
       id: source.id,
       length: source.length,
@@ -20,11 +21,10 @@ export const getShabad = async ( id: string ): Promise<ShabadData> => {
       id: writer.id,
       nameGurmukhi: writer.akhar,
     },
-    lines: shabad.map( ( { line: { id, translation, gurmukhi, linenum } } ) => ( {
+    lines: bani.map( ( { line: { id, translation, gurmukhi, lineno, pageno } } ) => ( {
       id,
       gurmukhi: gurmukhi.akhar,
-      sourceLine: linenum,
-      //! This is wrong for each line, but API doesn't provide this
+      sourceLine: lineno,
       sourcePage: pageno,
       //! translationSourceId is just set per language for now
       translations: [
