+++ src/services/data/collections/index.ts
@@ -1,4 +1,4 @@
-import { CollectionData } from '../../types/data'
+import { CollectionData } from '../../../types/data'
 import defaultCollections from './default'
 
 export const getCollections = (): Promise<CollectionData[]> => Promise.resolve( defaultCollections )
