+++ src/services/data/gurbaninow/types.ts
