+++ src/services/data/shabads.ts
@@ -1,16 +1,15 @@
-import Languages from '../lib/languages'
-import { BookmarkData } from '../types/data'
+import Languages from '../../lib/languages'
+import { ShabadData } from '../../types/data'
 import * as gurbaniNow from './gurbaninow'
 
-export const getBookmark = async ( id: string ): Promise<BookmarkData> => {
+export const getShabad = async ( id: string ): Promise<ShabadData> => {
   const {
-    baniinfo: { source, writer, akhar },
-    bani,
-  } = await gurbaniNow.getBani( id )
+    shabadinfo: { source, writer, pageno },
+    shabad,
+  } = await gurbaniNow.getShabad( id )
 
   return {
     id,
-    nameGurmukhi: akhar,
     source: {
       id: source.id,
       length: source.length,
@@ -21,10 +20,11 @@ export const getBookmark = async ( id: string ): Promise<BookmarkData> => {
       id: writer.id,
       nameGurmukhi: writer.akhar,
     },
-    lines: bani.map( ( { line: { id, translation, gurmukhi, lineno, pageno } } ) => ( {
+    lines: shabad.map( ( { line: { id, translation, gurmukhi, linenum } } ) => ( {
       id,
       gurmukhi: gurmukhi.akhar,
-      sourceLine: lineno,
+      sourceLine: linenum,
+      //! This is wrong for each line, but API doesn't provide this
       sourcePage: pageno,
       //! translationSourceId is just set per language for now
       translations: [
